OneShot: Add float calculation to HAL driver
Add the floating point calculation to pios driver. This makes the
firmware more cpu independent and allow the use of us-calculation
instead of fictional ranges.

+++ flight/PiOS/inc/pios_servo.h
@@ -34,7 +34,7 @@
 /* Public Functions */
 extern void PIOS_Servo_SetHz(const uint16_t * update_rates, uint8_t banks);
 extern void PIOS_Servo_Set(uint8_t Servo, uint16_t Position);
-extern void PIOS_Servo_OneShot_Set(uint8_t servo, uint16_t position);
+extern void PIOS_Servo_OneShot_Set(uint8_t servo, float position);
 extern void PIOS_Servo_OneShot_Update();
 
 #endif /* PIOS_SERVO_H */
